+++ src/components/common/filter/index.ts
@@ -1,3 +1,4 @@
+export { ApplicationSearchInput } from './ApplicationSearchInput'
 export { FilterBar } from './FilterBar'
 export { FilterSelect } from './FilterSelect'
 export { SearchInput } from './SearchInput'
