+++ src/features/recruitment/api/index.ts
@@ -1,0 +1,2 @@
+export * from './getAdminRecruitmentDetail'
+export * from './getAdminRecruitments'
