+++ src/features/recruitment/ui/detailModal/RecruitmentDetailContent.tsx
@@ -11,11 +11,11 @@ import type { ChangeEvent } from 'react'
 import { Link } from 'react-router'
 
 import { ApplicationStatusBadge } from '@/components/common/badge'
-import { getAdminRecruitmentDetail } from '@/features/recruitment/api/getAdminRecruitmentDetail'
-import { markdownToHtml } from '@/lib/markdown'
+import { getAdminRecruitmentDetail } from '@/features/recruitment/api'
+import markdownToHtml from '@/lib/markdown'
 import { tM } from '@/lib/twMerge'
-import { useRecruitmentDetailModalStore } from '@/store/recruitment/useRecruitmentModalStore'
-import { formatPrice } from '@/utils/price'
+import { useRecruitmentDetailModalStore } from '@/store/recruitment'
+import formatPrice from '@/utils/formatPrice'
 
 const LEFT_BOX_STYLE = 'flex flex-col gap-1 mb-4 cursor-default'
 const RIGHT_LEFT_BOX_STYLE = 'flex flex-col gap-1 mb-6 cursor-default'
