+++ src/store/recruitment/useRecruitmentModalStore.ts
@@ -7,14 +7,6 @@ type TagFilterModalState = {
   closeTagFilterModalModal: () => void
   toggleTagFilterModalModal: () => void
 }
-type DetailModalState = {
-  isDetailModalOpen: boolean
-  selectedRecruitmentId: number | null
-
-  openDetailModal: (value: number) => void
-  closeDetailModal: () => void
-  toggleDetailModal: () => void
-}
 
 export const useTagFilterModalStore = create<TagFilterModalState>((set) => ({
   isTagFilterModalOpen: false,
@@ -24,16 +16,3 @@ export const useTagFilterModalStore = create<TagFilterModalState>((set) => ({
   toggleTagFilterModalModal: () =>
     set((state) => ({ isTagFilterModalOpen: !state.isTagFilterModalOpen })),
 }))
-
-export const useRecruitmentDetailModalStore = create<DetailModalState>(
-  (set) => ({
-    isDetailModalOpen: false,
-    selectedRecruitmentId: null,
-
-    openDetailModal: (id) =>
-      set({ isDetailModalOpen: true, selectedRecruitmentId: id }),
-    closeDetailModal: () => set({ isDetailModalOpen: false }),
-    toggleDetailModal: () =>
-      set((state) => ({ isDetailModalOpen: !state.isDetailModalOpen })),
-  })
-)
